+++ CHANGELOG.md
@@ -11,6 +11,7 @@ All notable changes to this project will be documented in this file.
 - Option use_tuples which allows for mapping only overloaded procs to take one tuple, all procs to take tuples, skip mapping overloaded procs at all, and only mapping the first defined of the overloaded procs
 - Option for generated types to derive serde Serialize and Deserialize
 - Option for generating mapping as a whole crate directory
+- Option for changing output types, fields, and functions to use rust standard name casing (CamelCase for types and snake_case for fields and functions)
 
 ### Changed
 - Backend of code generation no long uses syn Node instead keeping everything in quote generated TokenStreams